combined forms into one with conditional fields
