+++ src/components/views/JobForm.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import { Field } from 'react-final-form';
 import { FileUploaderField } from '@folio/stripes-erm-components';
 import { AppIcon, TitleManager } from '@folio/stripes/core';
@@ -8,16 +9,22 @@ import stripesFinalForm from '@folio/stripes/final-form';
 
 import {
   Button,
+  Col,
   IconButton,
+  MessageBanner,
   Pane,
   PaneFooter,
   PaneMenu,
   Paneset,
+  Row,
+  TextField,
 } from '@folio/stripes/components';
+
+import { required } from '../../util/validators';
 
 import css from './JobForm.css';
 
-class JobForm extends React.Component {
+class JobFormKBART extends React.Component {
   static propTypes = {
     handleSubmit: PropTypes.func.isRequired,
     handlers: PropTypes.shape({
@@ -27,6 +34,7 @@ class JobForm extends React.Component {
     }),
     pristine: PropTypes.bool,
     submitting: PropTypes.bool,
+    format: PropTypes.string.isRequired
   }
 
   renderPaneFooter() {
@@ -86,7 +94,7 @@ class JobForm extends React.Component {
   }
 
   render() {
-    const { handlers: { onDownloadFile, onUploadFile } } = this.props;
+    const { handlers: { onDownloadFile, onUploadFile }, format } = this.props;
     return (
       <Paneset>
         <FormattedMessage id="ui-local-kb-admin.create">
@@ -97,11 +105,65 @@ class JobForm extends React.Component {
               footer={this.renderPaneFooter()}
               id="pane-job-form"
               firstMenu={this.renderFirstMenu()}
-              paneTitle={<FormattedMessage id="ui-local-kb-admin.job.newJSONJob" />}
+              paneTitle={<FormattedMessage id={`ui-local-kb-admin.job.new${format}Job`} />}
             >
               <TitleManager record={create}>
                 <form>
                   <div className={css.jobForm}>
+                    {format === 'KBART' &&
+                      <>
+                        <Row>
+                          <Col xs={12}>
+                            <MessageBanner>
+                              <SafeHTMLMessage id="ui-local-kb-admin.job.sourceReferenceWarning" />
+                            </MessageBanner>
+                          </Col>
+                        </Row>
+                        <Row>
+                          <Col xs={12}>
+                            <Field
+                              data-test-field-package-name
+                              component={TextField}
+                              name="packageName"
+                              label={<FormattedMessage id="ui-local-kb-admin.job.packageName" />}
+                              required
+                              validate={required}
+                            />
+                          </Col>
+                        </Row>
+                        <Row>
+                          <Col xs={4}>
+                            <Field
+                              data-test-field-package-source
+                              component={TextField}
+                              name="packageSource"
+                              label={<FormattedMessage id="ui-local-kb-admin.job.packageSource" />}
+                              required
+                              validate={required}
+                            />
+                          </Col>
+                          <Col xs={8}>
+                            <Field
+                              data-test-field-package-reference
+                              component={TextField}
+                              name="packageReference"
+                              label={<FormattedMessage id="ui-local-kb-admin.job.packageReference" />}
+                              required
+                              validate={required}
+                            />
+                          </Col>
+                        </Row>
+                        <Row>
+                          <Col xs={12}>
+                            <Field
+                              data-test-field-package-provider
+                              component={TextField}
+                              name="packageProvider"
+                              label={<FormattedMessage id="ui-local-kb-admin.job.packageProvider" />}
+                            />
+                          </Col>
+                        </Row>
+                      </> }
                     <Field
                       component={FileUploaderField}
                       data-test-document-field-file
@@ -110,6 +172,7 @@ class JobForm extends React.Component {
                       name="fileUpload"
                       onDownloadFile={onDownloadFile}
                       onUploadFile={onUploadFile}
+                      required
                       validate={this.validateUploadFile}
                     />
                   </div>
@@ -126,4 +189,4 @@ class JobForm extends React.Component {
 export default stripesFinalForm({
   navigationCheck: true,
   keepDirtyOnReinitialize: true,
-})(JobForm);
+})(JobFormKBART);
